backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-groq
chromadb
faiss-cpu
sentence-transformers
tiktoken
pymupdf
httpx
pandas
numpy
bcrypt
tavily-python
